Refactor '/aim/web/api' directory:  - Cleanup old not used views, models and utils  - Change paths for run structured data endpoints  - Minor fixes for endpoints after testing  - Minor fixes run.track() method  - Moved old utils that are still needed for storage files conversion to 'aim/web/legacy_api'
