--- aim/web/api/v2/helpers.py
@@ -1,11 +1,0 @@
-from fastapi import HTTPException
-
-from aim.web.api.projects.project import Project
-
-
-def object_factory():
-    project = Project()
-    if not project.exists():
-        raise HTTPException(status_code=404)
-
-    return project.repo.run_metadata_db
